+++ NEWS.md
@@ -1,5 +1,13 @@
 # Change Log
 
 ## Unreleased
-### Add
+### Added
+- depends and imports in DESCRIPTION
+- globalVariables for data frame variables in functions 
 - NEWS.md
+
+### Fixed
+- data documentation (wrong variable names)
+- missing parameter in `plot_splines`
+- missing `stats::` in `read_aaa`
+- `aes` call which caused notes on R CMD check using `aes_`